+++ modules/Shared/lang/en/messages.php
@@ -6,5 +6,8 @@ return [
     'record_created' => 'Record successfully created.',
     'record_updated' => 'Record successfully updated.',
     'record_saved' => 'Record successfully saved.',
+    'record_submitted' => 'Record successfully submitted.',
     'record_deleted' => 'Record successfully deleted.',
+    'record_approved' => 'Record successfully approved.',
+    'record_rejected' => 'Record successfully rejected.',
 ];